Fix delegates drawer
